+++ doc/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Current
 -------
 
+- Improve database connection handler
 - Change API endpoint for slots, coordinates are now handled in query parameters
 
 v1.0.0